+++ Examples/003-Arrays-In-Packets/ArraysInPackets.cs
@@ -8,7 +8,7 @@ namespace Link.Examples._003_Arrays_In_Packets;
 public static class ArraysInPackets
 {
     private const string IpAddress = "127.0.0.1";
-    private const int Port = 7777;
+    private const ushort Port = 7777;
     
     public static void Main()
     {
